Add tests for user service methods
This commit includes tests for the following methods in the user service:
- getAllUsers
- getUserById
- createUser
- updateUser
- getUserByResetToken
- updateUserPassword
- updateUserResetToken
- deleteUser

The tests cover both successful operation scenarios and cases where the methods are expected to return null. They use mock implementations of the UserRepository methods and check that the user service methods call the correct UserRepository methods with the correct arguments.

Adjustments have been made in the creation of ObjectId instances for _id fields in the mock data to ensure type compatibility with the IUser interface.

Tested scenarios include fetching users, creating a new user, updating user data and password, getting a user by reset token, updating a user's reset token, and deleting a user.